requirements.txt
django
ariadne
python-dateutil
